+++ src/genresmodel.cpp
@@ -72,7 +72,7 @@ namespace unplayer
 
     int GenresModel::rowCount(const QModelIndex&) const
     {
-        return mGenres.size();
+        return static_cast<int>(mGenres.size());
     }
 
     bool GenresModel::sortDescending() const
@@ -182,7 +182,7 @@ namespace unplayer
         }
         QObject::connect(LibraryUtils::instance(), &LibraryUtils::removingFilesChanged, this, [this, indexes] {
             if (!LibraryUtils::instance()->isRemovingFiles()) {
-                for (int i = indexes.size() - 1; i >= 0; --i) {
+                for (int i = static_cast<int>(indexes.size()) - 1; i >= 0; --i) {
                     const int index = indexes[i];
                     beginRemoveRows(QModelIndex(), index, index);
                     mGenres.erase(mGenres.begin() + index);
